# docs/api-reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "C-Address Onboarding Bridge API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Bridge API",
      "url": "https://github.com/C-Address-Onboarding-Bridge"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for authenticating requests. Obtainable via POST /api/v1/keys."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "validation_error"
          },
          "message": {
            "type": "string",
            "example": "invalid request"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "path": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "path",
                "message"
              ]
            }
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
      "TxStatusResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "success",
              "failed"
            ]
          },
          "hash": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$",
            "description": "64-char hex transaction hash",
            "example": "a1b2c3d4..."
          },
          "explorerUrl": {
            "type": "string",
            "format": "uri"
          },
          "explorerUrls": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uri"
            }
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "status",
          "hash"
        ],
        "title": "TxStatusResponse"
      },
...
      "FundRequest": {
        "type": "object",
        "properties": {
          "signedXdr": {
            "type": "string",
            "minLength": 1,
            "description": "Base64-encoded signed Stellar transaction XDR",
            "example": "AAAAAgAAAABxy..."
          }
        },
        "required": [
          "signedXdr"
        ],
        "title": "FundRequest"
      },
      "FundPrepareRequest": {
        "type": "object",
        "properties": {
          "sourceAddress": {
            "type": "string",
            "pattern": "^[GC][A-Z2-7]{55}$",
            "description": "Stellar/C-Address (56-char base32)",
            "example": "GABCDE...WXYZ"
          },
          "targetAddress": {
            "type": "string",
            "pattern": "^[GC][A-Z2-7]{55}$",
            "description": "Stellar/C-Address (56-char base32)",
            "example": "GABCDE...WXYZ"
          },
          "tokenAddress": {
            "type": "string",
            "pattern": "^[GC][A-Z2-7]{55}$",
            "description": "Token contract address",
            "example": "GABCDE...WXYZ"
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Amount in stroops (1 XLM = 10,000,000 stroops)",
            "example": "1000000"
          },
          "memo": {
            "type": "string",
            "maxLength": 64,
            "example": "bridge-payment"
          }
        },
        "required": [
          "sourceAddress",
          "targetAddress",
          "tokenAddress",
          "amount"
        ],
        "title": "FundPrepareRequest"
      },
...
      "CexRouteRequest": {
        "type": "object",
        "properties": {
          "exchange": {
            "type": "string",
            "enum": [
              "binance",
              "coinbase",
              "kraken",
              "generic"
            ],
            "example": "binance"
          },
          "sourceAsset": {
            "type": "string",
            "minLength": 1,
            "example": "XLM"
          },
          "amount": {
            "type": "string",
            "pattern": "^\\d+$",
            "description": "Amount in stroops (1 XLM = 10,000,000 stroops)",
            "example": "1000000"
          },
          "targetCAddress": {
            "type": "string",
            "pattern": "^[GC][A-Z2-7]{55}$",
            "description": "Stellar/C-Address (56-char base32)",
            "example": "GABCDE...WXYZ"
          },
          "targetNetwork": {
            "type": "string",
            "default": "stellar"
          },
          "memo": {
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "exchange",
          "sourceAsset",
          "amount",
          "targetCAddress"
        ],
        "title": "CexRouteRequest"
      },
...
    },
...
  },
  "paths": {
...
    "/api/v2/fund": {
      "post": {
        "operationId": "submitFunding",
        "summary": "Submit a signed funding transaction",
        "description": "Submits a pre-signed Stellar transaction XDR to the Soroban bridge contract. Use POST /api/v2/fund/prepare to build the transaction first.",
        "tags": [
          "Fund"
        ],
        "security": [
          {
            "ApiKeyAuth": [
              "fund:write"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxStatusResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid XDR or validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Soroban RPC error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/fund/prepare": {
      "post": {
        "operationId": "prepareFunding",
        "summary": "Prepare an unsigned funding transaction for signing",
        "description": "Simulates the contract call and returns an unsigned transaction XDR for the client to sign. Sign this with your wallet and POST it to /api/v2/fund.",
        "tags": [
          "Fund"
        ],
        "security": [
          {
            "ApiKeyAuth": [
              "fund:write"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundPrepareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Unsigned transaction ready for signing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "instruction": {
                      "type": "string"
                    },
                    "simulation": {},
                    "params": {
                      "$ref": "#/components/schemas/FundPrepareRequest"
                    }
                  },
                  "required": [
                    "instruction",
                    "params"
                  ],
                  "title": "FundPrepareResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/cex/route": {
      "post": {
        "operationId": "cexRoute",
        "summary": "Get CEX withdrawal routing instructions",
        "description": "Returns withdrawal configuration for routing from a supported CEX to a C-Address.",
        "tags": [
          "CEX"
        ],
        "security": [
          {
            "ApiKeyAuth": [
              "cex:read"
            ]
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CexRouteRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "CEX withdrawal route",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "exchange": {
                      "type": "string"
                    },
                    "withdrawalAddress": {
                      "type": "string"
                    },
                    "memo": {
                      "type": "string"
                    },
                    "network": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "exchange",
                    "withdrawalAddress",
                    "network",
                    "amount"
                  ],
                  "title": "CexRouteResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}